One hot encode on the fly, Save pkls throughout load_data

+++ trainer/util.py
@@ -103,22 +103,36 @@ def build_ingr_id_dict(ingr_ids, id2rvocab):
         rid2iid[rid] = iid
     return id2ingr, rid2iid
 
-def encode_ingrs(ingr_ids_list, rid2iid):
+def reindex_ingr_ids(ingr_ids_list, rid2iid):
     """
-    One hot encode array of raw ingredient labels.
+    Re-index ingredient ids off recipe vocab and onto ingredient only vocab
+    for one hot encoding.
     """
+    new_list = []
 
-    ningrs = len(list(rid2iid))
-    nsamples = len(ingr_ids_list)
-    one_hot = np.zeros((nsamples, ningrs))
-    
     for i, ingr_ids in enumerate(ingr_ids_list):
+        new_IDs = []
         for rid in ingr_ids:
             if rid == 1: # stop token
                 break
             iid = rid2iid[rid]
             if iid != -1: # checks if id found in dict (and thus a valid ingr)
-                one_hot[i][iid] = 1
+                new_IDs.append(iid)
+        new_list.append(new_IDs)
+    
+    return new_list
+
+def encode_ingrs(ingr_ids_list, ningrs):
+    """
+    One hot encode array of ingredients-only indexed ids.
+    """
+
+    nsamples = len(ingr_ids_list)
+    one_hot = np.zeros((nsamples, ningrs))
+    
+    for i, ingr_ids in enumerate(ingr_ids_list):
+        for iid in ingr_ids:
+            one_hot[i][iid] = 1
     return one_hot
 
 ### Read LMDB into raw labels dataset
@@ -175,7 +189,6 @@ def read_lmdb(lmdb_path):
 
 def load_data(dirs, files):
     """
-    TODO: add test partition, details in code
     Processes and returns all splits of data from the prespecified data directories.
     """
 
@@ -189,23 +202,33 @@ def load_data(dirs, files):
     if not (os.path.isfile(lmdb_tr_out_path) and os.path.isfile(lmdb_val_out_path) and os.path.isfile(lmdb_test_out_path)):
         print("Error: one of the LMDB output files does not exist. Try re-reading the LMDBs.")
         return
-    
-    id2raw_labels_tr = pickle.load(open(lmdb_tr_out_path, 'rb'))
-    id2raw_labels_val = pickle.load(open(lmdb_val_out_path, 'rb'))
-    id2raw_labels_test = pickle.load(open(lmdb_test_out_path, 'rb'))
+    else:
+        id2raw_labels_tr = pickle.load(open(lmdb_tr_out_path, 'rb'))
+        id2raw_labels_val = pickle.load(open(lmdb_val_out_path, 'rb'))
+        id2raw_labels_test = pickle.load(open(lmdb_test_out_path, 'rb'))
+        print("Raw data from LMDBs read in.")
     
     # Save valid image ids
-    img_ids_tr = get_valid_ids(img_dir_tr, id2raw_labels_tr)
-    img_ids_val = get_valid_ids(img_dir_val, id2raw_labels_val)
-    img_ids_test = get_valid_ids(img_dir_test, id2raw_labels_test)
-    
-    partition = {
-        'train': img_ids_tr, 
-        'validation': img_ids_val,
-        'test': img_ids_test,
-    }
-    pickle.dump(partition, open(files['pt_pkl'], "wb"))
-    print("Valid image ids dict saved as pkl.")
+    if os.path.isfile(files['pt_pkl']):
+        with open(files['pt_pkl'], 'rb') as pt:
+            partition = pickle.load(pt)
+        img_ids_tr = partition['train']
+        img_ids_val = partition['validation']
+        img_ids_test = partition['test']
+        print("Valid image ids loaded from pkl")
+    else:
+        subdirect = False # for testing
+        img_ids_tr = get_valid_ids(img_dir_tr, id2raw_labels_tr, subdirect = subdirect)
+        img_ids_val = get_valid_ids(img_dir_val, id2raw_labels_val, subdirect = subdirect)
+        img_ids_test = get_valid_ids(img_dir_test, id2raw_labels_test, subdirect = subdirect)
+        
+        partition = {
+            'train': img_ids_tr, 
+            'validation': img_ids_val,
+            'test': img_ids_test,
+        }
+        pickle.dump(partition, open(files['pt_pkl'], "wb"))
+        print("Valid image ids dict saved as pkl.")
     
     # Store the corresponding raw labels
     raw_labels_tr = np.array([id2raw_labels_tr[i] for i in img_ids_tr])
@@ -225,37 +248,50 @@ def load_data(dirs, files):
         id2class = pickle.load(f) # class_id to class_name dict
     
     # TESTING: don't one hot encode to save memory
-#     # One hot encode class ids
-#     class_labels = encode_class(class_ids, id2class)
+    # # One hot encode class ids
+    # class_labels = encode_class(class_ids, id2class)
     
     # Build ingr id dictionary
-    with open(recipe_vocab_path) as f_vocab:
-        id2rvocab = {i+2: w.rstrip() for i, w in enumerate(f_vocab)}
-        id2rvocab[1] = '</i>'
-    rvocab2id = {v:k for k,v in id2rvocab.items()}
+    if os.path.isfile(files['ingr_pkl']) and os.path.isfile(files['ingr_id2id_pkl']):
+        with open(files['ingr_pkl'], 'rb') as ingr_pkl:
+            id2ingr = pickle.load(ingr_pkl)
+        with open(files['ingr_id2id_pkl'], 'rb') as ingr_id2id_pkl:
+            rid2iid = pickle.load(ingr_id2id_pkl)
+        print("Ingredients name and old-to-new id dicts loaded from pkl.")
+    else:
+        with open(recipe_vocab_path) as f_vocab:
+            id2rvocab = {i+2: w.rstrip() for i, w in enumerate(f_vocab)}
+            id2rvocab[1] = '</i>'
+        rvocab2id = {v:k for k,v in id2rvocab.items()}
+        
+        invalid_ingr_names = ['Ingredients', '1', '100', '2', '200', '23', '30', '300', \
+                            '4', '450', '50', '500', '6', '600']
+        invalid_ids = [rvocab2id[n] for n in invalid_ingr_names]
+        id_ds_list = [id2raw_labels_tr, id2raw_labels_val, id2raw_labels_test]
+        unique_ids = find_unique_ingr_ids(id_ds_list, invalid_ids)
+        id2ingr, rid2iid = build_ingr_id_dict(unique_ids, id2rvocab)
+        print("# unique ingredients:", len(list(id2ingr)))
+        
+        # Save ingr id and rid2iid dictionary
+        pickle.dump(id2ingr, open(files['ingr_pkl'], "wb"))
+        pickle.dump(rid2iid, open(files['ingr_id2id_pkl'], "wb"))
+        print("Ingredients name and old-to-new id dicts saved as pkl.")
     
-    invalid_ingr_names = ['Ingredients', '1', '100', '2', '200', '23', '30', '300', \
-                          '4', '450', '50', '500', '6', '600']
-    invalid_ids = [rvocab2id[n] for n in invalid_ingr_names]
-    id_ds_list = [id2raw_labels_tr, id2raw_labels_val, id2raw_labels_test]
-    unique_ids = find_unique_ingr_ids(id_ds_list, invalid_ids)
-    id2ingr, rid2iid = build_ingr_id_dict(unique_ids, id2rvocab)
-    print("# unique ingredients:", len(list(id2ingr)))
-    
-    # Save ingr id dictionary
-    pickle.dump(id2ingr, open(files['ingr_pkl'], "wb"))
-    print("Ingredients name dict saved as pkl.")
-    
+    # TESTING: don't one hot encode to save memory
     # One hot encode ingr ids
-    ingr_labels = encode_ingrs(ingr_ids, rid2iid)
+    ingr_ids = reindex_ingr_ids(ingr_ids, rid2iid)
+    ingr_labels = encode_ingrs(ingr_ids, len(id2ingr))
     
     # Create new labels dictionary
-    ingr_gb = getsizeof(ingr_labels)/(2**30)
+    ingr_gb = getsizeof(ingr_ids)/(2**30) 
     class_gb = getsizeof(class_ids)/(2**30)
     nsteps_gb = getsizeof(nsteps)/(2**30)
-    print('Size of Ingr, Class, Nsteps labels: {:.5f} {:.5f} {:.5f} GB'.format(ingr_gb, class_gb, nsteps_gb))
+    print('Size of Ingr, Class, Nsteps labels: {:.8f} {:.5f} {:.5f} GB'.format(ingr_gb, class_gb, nsteps_gb))
 
-    labels = dict(zip(img_ids, zip(ingr_labels, class_ids, nsteps)))
+    ingr_enc_gb = getsizeof(ingr_labels)/(2**30)
+    print('Size of Ingr one hot encoded labels: {:.5f} GB'.format(ingr_enc_gb))
+
+    labels = dict(zip(img_ids, zip(ingr_ids, class_ids, nsteps))) # ingr_labels
     pickle.dump(labels, open(files['lb_pkl'], "wb"))
     print("Labels dict saved as pkl.")
